New version

+++ Feduino_with_webcontrol_beta/Grava_Sd_card.ino
@@ -7,19 +7,15 @@ void writeCRLF(SdFile& f)
 
 void logtempgraf() //Grava dados no SD CARD para gerar  grafico de temperatura.
 {
-  char oldminuto1;
-  oldminuto1 = time1;
-  time1 = t.min;
-
   if(NumMins(t.hour,t.min) == 1)
   {
     EEPROM.write(688, t.date); 
   }
 
-  if (oldminuto1!=time1)
-  {
-    file.open(&root, "LOGTDIA.TXT", O_CREAT | O_APPEND | O_WRITE);
+  selecionar_SPI(SD_CARD);
 
+  if(file.open(&root, "LOGTDIA.TXT", O_CREAT | O_APPEND | O_WRITE))
+  {
     if ((tempC<=1.0) || (tempC>99.9))
     {
       file.print("00.00");
@@ -32,28 +28,29 @@ void logtempgraf() //Grava dados no SD CARD para gerar  grafico de temperatura.
       file.write((uint8_t*)"\0", 1);
       writeCRLF(file);
     }
+    file.close();
   }
-  file.close();
 
   if ((t.hour == 23) && (t.min > 58))
   {
-    file.open(&root, "LOGTDIA.TXT", O_WRITE);
-    file.remove();
-  } 
+    if(file.open(&root, "LOGTDIA.TXT", O_WRITE))
+    {
+      file.remove();
+      file.close();
+    }
+  }
 }
 
 void logphagraf() //Grava dados no SD CARD para gerar  gráfico de PH do aquário.
 { 
-  char oldminuto2;
-  oldminuto2 = time2;
-  time2 = t.min;
-
   if(NumMins(t.hour,t.min) == 1)
   {
     EEPROM.write(689, t.date); 
   }
 
-  if (oldminuto2!=time2)
+  selecionar_SPI(SD_CARD);
+  
+  if(file.open(&root, "LOGPDIA.TXT", O_CREAT | O_APPEND | O_WRITE))
   {
     if((PHA < 1.00) || (PHA > 9.99))
     {
@@ -62,32 +59,33 @@ void logphagraf() //Grava dados no SD CARD para gerar  gráfico de PH do aquári
       writeCRLF(file);
     }
     else
-    {
-      file.open(&root, "LOGPDIA.TXT", O_CREAT | O_APPEND | O_WRITE);   
+    {   
       file.print(PHA);
       file.write((uint8_t*)"\0", 1);
       writeCRLF(file);
     }
+    file.close();
   }
-  file.close();
-
   if ((t.hour == 23) && (t.min > 58))
   {
-    file.open(&root, "LOGPDIA.TXT", O_WRITE);
-    file.remove();
+    if(file.open(&root, "LOGPDIA.TXT", O_WRITE))
+    {
+      file.remove();
+      file.close();
+    }
   } 
 }
 
 void logphrgraf() //Grava dados no SD CARD para gerar  gráfico de PH do reator de calcio.
 {
-  char oldminuto3;
-  oldminuto3 = time3;
-  time3 = t.min;
   if(NumMins(t.hour,t.min) == 1)
   {
     EEPROM.write(690, t.date); 
   }
-  if (oldminuto3!=time3)
+
+  selecionar_SPI(SD_CARD);
+  
+  if(file.open(&root, "LOGRDIA.TXT", O_CREAT | O_APPEND | O_WRITE))
   {
     if((PHR < 1.00) || (PHR > 9.99))
     {
@@ -97,34 +95,34 @@ void logphrgraf() //Grava dados no SD CARD para gerar  gráfico de PH do reator 
     }
     else
     {
-      file.open(&root, "LOGRDIA.TXT", O_CREAT | O_APPEND | O_WRITE);   
       file.print(PHR);
       file.write((uint8_t*)"\0", 1);
       writeCRLF(file);
     }
+    file.close();
   }
-  file.close();
 
   if ((t.hour == 23) && (t.min > 58))
   {
-    file.open(&root, "LOGRDIA.TXT", O_WRITE);
-    file.remove();
+    if(file.open(&root, "LOGRDIA.TXT", O_WRITE))
+    {
+      file.remove();
+      file.close();
+    }
   }
 }
 
 void logorpgraf() //Grava dados no SD CARD para gerar  grafico de ORP.
 {
-  char oldminuto4;
-  oldminuto4 = time4;
-  time4 = t.min;
-
   if(NumMins(t.hour,t.min) == 1)
   {
     EEPROM.write(691, t.date); 
   }
 
-  if (oldminuto4!=time4)
-  {
+  selecionar_SPI(SD_CARD);
+  
+  if(file.open(&root, "LOGODIA.TXT", O_CREAT | O_APPEND | O_WRITE))
+  {   
     if ((ORP<=100) || (tempC>999))
     {
       file.print("000");
@@ -133,33 +131,33 @@ void logorpgraf() //Grava dados no SD CARD para gerar  grafico de ORP.
     }
     else
     {
-      file.open(&root, "LOGODIA.TXT", O_CREAT | O_APPEND | O_WRITE);   
       file.print(ORP);
       file.write((uint8_t*)"\0", 1);
       writeCRLF(file);
     }
+    file.close();
   }
-  file.close();
 
   if ((t.hour == 23) && (t.min > 58))
   {
-    file.open(&root, "LOGODIA.TXT", O_WRITE);
-    file.remove();
+    if(file.open(&root, "LOGODIA.TXT", O_WRITE))
+    {
+      file.remove();
+      file.close();
+    }
   } 
 }
 
 void logdengraf() //Grava dados no SD CARD para gerar  grafico de PH do aquario.
 {
-  char oldminuto5;
-  oldminuto5 = time5;
-  time5  = t.min;
-
   if(NumMins(t.hour,t.min) == 1)
   {
     EEPROM.write(692, t.date); 
   }
 
-  if (oldminuto5!=time5)
+  selecionar_SPI(SD_CARD);
+  
+  if(file.open(&root, "LOGDDIA.TXT", O_CREAT | O_APPEND | O_WRITE))   
   {
     if( (DEN<1000) || (DEN>9999))
     {
@@ -169,7 +167,7 @@ void logdengraf() //Grava dados no SD CARD para gerar  grafico de PH do aquario.
     }
     else
     {    
-      file.open(&root, "LOGDDIA.TXT", O_CREAT | O_APPEND | O_WRITE);   
+
       file.print(DEN);
       file.write((uint8_t*)"\0", 1);
       writeCRLF(file);
@@ -179,25 +177,28 @@ void logdengraf() //Grava dados no SD CARD para gerar  grafico de PH do aquario.
 
   if ((t.hour == 23) && (t.min > 58))
   {
-    file.open(&root, "LOGDDIA.TXT", O_WRITE);
-    file.remove();
-  }  
+    if(file.open(&root, "LOGDDIA.TXT", O_WRITE))
+    {
+      file.remove();
+      file.close();
+    }
+  } 
 }
-void check_arquivo_temp_agua(boolean web = false, int num = 0)
+
+void check_arquivo_temp_agua()
 {
   int16_t n;
   char buf[8];
   int contador = 0;
-  EthernetClient client = server.available();
+
   byte k = EEPROM.read(688);
 
-  if(web == false)
+  selecionar_SPI(SD_CARD);
+  
+  if(k != t.date)
   {
-    if(k != t.date)
-    {
-      file.open(&root, "LOGTDIA.TXT", O_READ);
-      file.remove();
-    }
+    file.open(&root, "LOGTDIA.TXT", O_WRITE);
+    file.remove();
   }
 
   file.open(&root, "LOGTDIA.TXT", O_CREAT | O_READ | O_APPEND | O_WRITE);
@@ -207,62 +208,33 @@ void check_arquivo_temp_agua(boolean web = false, int num = 0)
     {
       contador++;
     }
-    if(web == true)
-    {     
-      if(contador == (num + 1))
-      {
-        client.print("{\"v0\":");
-      } 
-      if((contador > num) && (contador < num + 15))
-      {
-        client.print(buf);
-        client.print(",\"v" + String(contador - num) + "\":");
-      }
-      if(contador == (num + 15))
-      {
-        client.print(buf);  
-      }
-    }
   }
-  if(web == true)
+  if(contador < NumMins(t.hour, t.min))
   {
-    if((num + 15) > contador)
+    for(int i = contador; i < NumMins(t.hour, t.min) - 1; i++)
     {
-      client.print("\"\""); 
-    }
-    client.print("}"); 
-  }
-
-  if(web == false)
-  {
-    if(contador < NumMins(t.hour, t.min))
-    {
-      for(int i = contador; i < NumMins(t.hour, t.min) - 1; i++)
-      {
-        file.print("00.00");
-        file.write((uint8_t*)"\0", 1);
-        writeCRLF(file);
-      } 
-    }
+      file.print("00.00");
+      file.write((uint8_t*)"\0", 1);
+      writeCRLF(file);
+    } 
   }
   file.close();
 }
 
-void check_arquivo_ph_agua(boolean web = false, int num = 0)
+void check_arquivo_ph_agua()
 {
   int16_t n;
   char buf[7];
   int contador = 0;
-  EthernetClient client = server.available();
+
   byte k = EEPROM.read(689);
 
-  if(web == false)
+  selecionar_SPI(SD_CARD);
+  
+  if(k != t.date)
   {
-    if(k != t.date)
-    {
-      file.open(&root, "LOGPDIA.TXT", O_READ);
-      file.remove();
-    }
+    file.open(&root, "LOGPDIA.TXT", O_WRITE);
+    file.remove();
   }
 
   file.open(&root, "LOGPDIA.TXT", O_CREAT | O_READ | O_APPEND | O_WRITE);
@@ -272,61 +244,33 @@ void check_arquivo_ph_agua(boolean web = false, int num = 0)
     {
       contador++;
     }
-    if(web == true)
-    {     
-      if(contador == (num + 1))
-      {
-        client.print("{\"v0\":");
-      } 
-      if((contador > num) && (contador < num + 15))
-      {
-        client.print(buf);
-        client.print(",\"v" + String(contador - num) + "\":");
-      }
-      if(contador == (num + 15))
-      {
-        client.print(buf);  
-      }
-    }
   }
-  if(web == true)
+  if(contador < NumMins(t.hour, t.min))
   {
-    if((num + 15) > contador)
+    for(int i = contador; i < NumMins(t.hour, t.min) - 1; i++)
     {
-      client.print("\"\""); 
-    }
-    client.print("}"); 
-  }
-  if(web == false)
-  {
-    if(contador < NumMins(t.hour, t.min))
-    {
-      for(int i = contador; i < NumMins(t.hour, t.min) - 1; i++)
-      {
-        file.print("0.00");
-        file.write((uint8_t*)"\0", 1);
-        writeCRLF(file);
-      } 
-    }
+      file.print("0.00");
+      file.write((uint8_t*)"\0", 1);
+      writeCRLF(file);
+    } 
   }
   file.close();
 }
 
-void check_arquivo_ph_reator(boolean web = false, int num = 0)
+void check_arquivo_ph_reator()
 {
   int16_t n;
   char buf[7];
   int contador = 0;
-  EthernetClient client = server.available();
+
   byte k = EEPROM.read(690);
 
-  if(web == false)
+  selecionar_SPI(SD_CARD);
+  
+  if(k != t.date)
   {
-    if(k != t.date)
-    {
-      file.open(&root, "LOGRDIA.TXT", O_READ);
-      file.remove();
-    }
+    file.open(&root, "LOGRDIA.TXT", O_WRITE);
+    file.remove();
   }
 
   file.open(&root, "LOGRDIA.TXT", O_CREAT | O_READ | O_APPEND | O_WRITE);
@@ -336,61 +280,33 @@ void check_arquivo_ph_reator(boolean web = false, int num = 0)
     {
       contador++;
     }
-    if(web == true)
-    {     
-      if(contador == (num + 1))
-      {
-        client.print("{\"v0\":");
-      } 
-      if((contador > num) && (contador < num + 15))
-      {
-        client.print(buf);
-        client.print(",\"v" + String(contador - num) + "\":");
-      }
-      if(contador == (num + 15))
-      {
-        client.print(buf);  
-      }
-    }
   }
-  if(web == true)
+  if(contador < NumMins(t.hour, t.min))
   {
-    if((num + 15) > contador)
+    for(int i = contador; i < NumMins(t.hour, t.min) - 1; i++)
     {
-      client.print("\"\""); 
-    }
-    client.print("}"); 
-  }
-  if(web == false)
-  {
-    if(contador < NumMins(t.hour, t.min))
-    {
-      for(int i = contador; i < NumMins(t.hour, t.min) - 1; i++)
-      {
-        file.print("0.00");
-        file.write((uint8_t*)"\0", 1);
-        writeCRLF(file);
-      } 
-    }
+      file.print("0.00");
+      file.write((uint8_t*)"\0", 1);
+      writeCRLF(file);
+    } 
   }
   file.close();
 }
 
-void check_arquivo_orp(boolean web = false, int num = 0)
+void check_arquivo_orp()
 {
   int16_t n;
   char buf[6];
   int contador = 0;
-  EthernetClient client = server.available();
+
   byte k = EEPROM.read(691);
 
-  if(web == false)
+  selecionar_SPI(SD_CARD);
+  
+  if(k != t.date)
   {
-    if(k != t.date)
-    {
-      file.open(&root, "LOGODIA.TXT", O_READ);
-      file.remove();
-    }
+    file.open(&root, "LOGODIA.TXT", O_WRITE);
+    file.remove();
   }
 
   file.open(&root, "LOGODIA.TXT", O_CREAT | O_READ | O_APPEND | O_WRITE);
@@ -400,61 +316,33 @@ void check_arquivo_orp(boolean web = false, int num = 0)
     {
       contador++;
     }
-    if(web == true)
-    {     
-      if(contador == (num + 1))
-      {
-        client.print("{\"v0\":");
-      } 
-      if((contador > num) && (contador < num + 15))
-      {
-        client.print(buf);
-        client.print(",\"v" + String(contador - num) + "\":");
-      }
-      if(contador == (num + 15))
-      {
-        client.print(buf);  
-      }
-    }
   }
-  if(web == true)
+  if(contador < NumMins(t.hour, t.min))
   {
-    if((num + 15) > contador)
+    for(int i = contador; i < NumMins(t.hour, t.min) - 1; i++)
     {
-      client.print("\"\""); 
-    }
-    client.print("}"); 
-  }
-  if(web == false)
-  {
-    if(contador < NumMins(t.hour, t.min))
-    {
-      for(int i = contador; i < NumMins(t.hour, t.min) - 1; i++)
-      {
-        file.print("000");
-        file.write((uint8_t*)"\0", 1);
-        writeCRLF(file);
-      } 
-    }
+      file.print("000");
+      file.write((uint8_t*)"\0", 1);
+      writeCRLF(file);
+    } 
   }
   file.close();
 }
 
-void check_arquivo_densidade(boolean web = false, int num = 0)
+void check_arquivo_densidade()
 {
   int16_t n;
   char buf[7];
   int contador = 0;
-  EthernetClient client = server.available();
+
   byte k = EEPROM.read(692);
 
-  if(web == false)
+  selecionar_SPI(SD_CARD);
+  
+  if(k != t.date)
   {
-    if(k != t.date)
-    {
-      file.open(&root, "LOGDDIA.TXT", O_READ);
-      file.remove();
-    }
+    file.open(&root, "LOGDDIA.TXT", O_WRITE);
+    file.remove();
   }
 
   file.open(&root, "LOGDDIA.TXT", O_CREAT | O_READ | O_APPEND | O_WRITE);
@@ -464,43 +352,16 @@ void check_arquivo_densidade(boolean web = false, int num = 0)
     {
       contador++;
     }
-    if(web == true)
-    {     
-      if(contador == (num + 1))
-      {
-        client.print("{\"v0\":");
-      } 
-      if((contador > num) && (contador < num + 15))
-      {
-        client.print(buf);
-        client.print(",\"v" + String(contador - num) + "\":");
-      }
-      if(contador == (num + 15))
-      {
-        client.print(buf);  
-      }
-    }
   }
-  if(web == true)
+
+  if(contador < NumMins(t.hour, t.min))
   {
-    if((num + 15) > contador)
+    for(int i = contador; i < NumMins(t.hour, t.min) - 1; i++)
     {
-      client.print("\"\""); 
-    }
-    client.print("}"); 
-  }
-  if(web == false)
-  {
-    if(contador < NumMins(t.hour, t.min))
-    {
-      for(int i = contador; i < NumMins(t.hour, t.min) - 1; i++)
-      {
-        file.print("0000");
-        file.write((uint8_t*)"\0", 1);
-        writeCRLF(file);
-      } 
-    }
+      file.print("0000");
+      file.write((uint8_t*)"\0", 1);
+      writeCRLF(file);
+    } 
   }
   file.close();
 }
-
